requirements.txt
pillow==7.2.0 
qrcode==6.1 
six==1.15.0